Guard audio load slot polling

+++ src/main/PAL/main/audio/vab_sequence_slots.c
@@ -21,7 +21,7 @@ s32 OpenSequenceAudioSlot(u8 *header, u8 *body, void *seq) {
     }
     g_SoundScale.vabIds[AUDIO_SLOT_SEQUENCE] = vabId;
 
-    g_SeqHandle.storage = (s16)SsSeqOpen(seq, vabId);
+    g_SeqHandle.storage = SsSeqOpen(seq, vabId);
     g_SeqVolumeFadeStep = 0;
     g_VabTransferDone = SsVabTransCompleted(0);
     return g_VabTransferDone;
